requirements.txt
Django==5.0.9
djangorestframework
channels[daphne]
psycopg[binary]
python-dotenv
channels_redis
django-cors-headers
django-apscheduler
httpx
djangorestframework-simplejwt